perf(zkevm): keep the stack pointer and code length at native width
Measured on prover cost: total falls 0.73%, of which MEMORY is 5.31%.

`ziskemu -X` charges 122 for a four-byte field read and 193 for a write against
16 and 18 at native width - 7.6x and 10.7x - and none of that shows in the step
count, where `lw` and `sw` are one step each. `EvmStack` lays out `_tracer`(0),
`_stack`(8), `Code`(16), `Head`(24), `CodeLength`(28), and the last two are the
hottest four-byte fields in the guest: Head is read and written per stack
operation, CodeLength read per dispatch.

Storing them wide is only half of it. Reading Head back through an int-typed
property lets the compiler fold the load to a four-byte one - equivalent on a
little-endian target - so the writes widen and the reads do not. The offset
therefore stays native-width from the field to the point of use, which also drops
the widening each slot address needed: `(nint)(headOffset * WordSize)` had been
sign-extending a 32-bit product on every push and pop.

`ReadEip8024ImmediateOrZero` takes its length as nint too, removing a conversion
against an already-nint program counter on both targets.

+++ src/Nethermind/Nethermind.Evm/Instructions/EvmInstructions.CodeCopy.cs
@@ -72,7 +72,7 @@ public static partial class EvmInstructions
         where TGasPolicy : struct, IGasPolicy<TGasPolicy>
         where TTracingInst : struct, IFlag
         => DataCopy<TGasPolicy, TTracingInst>(vm, ref stack, ref gas,
-            MemoryMarshal.CreateReadOnlySpan(ref stack.Code, stack.CodeLength));
+            MemoryMarshal.CreateReadOnlySpan(ref stack.Code, (int)stack.CodeLength));
 
     /// <summary>
     /// CALLDATACOPY - copies a portion of the transaction's calldata into memory.
